Library

Library

Thunks module

Jobs module

Thunks module

EasyJobs.Thunks.Thunk — Type

Thunk(::Function, args::Tuple, kwargs::NamedTuple)
Thunk(::Function, args...; kwargs...)
Thunk(::Function)

Hold a Function and its arguments for lazy evaluation. Use reify! to evaluate.

Examples

julia> using EasyJobs.Thunks

julia> a = Thunk(x -> 3x, 4);

julia> reify!(a)
Some(12)

julia> b = Thunk(+, 4, 5);

julia> reify!(b)
Some(9)

julia> c = Thunk(sleep)(1);

julia> getresult(c)  # `c` has not been evaluated

julia> reify!(c)  # `c` has been evaluated
Some(nothing)

julia> f(args...; kwargs...) = collect(kwargs);

julia> d = Thunk(f)(1, 2, 3; x=1.0, y=4, z="5");

julia> reify!(d)
Some(Pair{Symbol, Any}[:x => 1.0, :y => 4, :z => "5"])

julia> e = Thunk(sin, "1");  # Catch errors

julia> reify!(e);

EasyJobs.Thunks.ErredResult — Type

Capture errors and stack traces from a running Thunk.

EasyJobs.Thunks.reify! — Function

reify!(thunk::Thunk)

Reify a Thunk into a value.

Compute the value of the expression. Walk through the Thunk's arguments and keywords, recursively evaluating each one, and then evaluating the Thunk's function with the evaluated arguments.

See also Thunk.

EasyJobs.Thunks.getresult — Method

getresult(thunk::Thunk)

Get the result of a Thunk. If thunk has not been evaluated, return nothing, else return a Some-wrapped result.

Jobs module

EasyJobs.SimpleJob — Type

SimpleJob(core::Thunk; description="", username="")

Create a simple job.

Arguments

core: a Thunk that encloses the job definition.

= name: give a short name to the job.

description::String="": describe what the job does in more detail.

username::String="": indicate who executes the job.

Examples

julia> using EasyJobs.Thunks

julia> a = SimpleJob(Thunk(sleep)(5); user="me", description="Sleep for 5 seconds");

julia> b = SimpleJob(Thunk(run, `pwd` & `ls`); username="me", description="Run some commands");

EasyJobs.Thunks.getresult — Method

getresult(job::Job)

Get the running result of a Job.

The result is wrapped by a Some type. Use something to retrieve its value. If it is nothing, the Job is not finished.

EasyJobs.getstatus — Method

getstatus(x::Job)

Get the current status of a Job.

EasyJobs.ispending — Function

Test if the Job is still pending.

EasyJobs.isrunning — Function

Test if the Job is running.

EasyJobs.isexited — Function

Test if the Job has exited.

EasyJobs.issucceeded — Function

Test if the Job was successfully run.

EasyJobs.isfailed — Function

Test if the Job failed during running.

EasyJobs.isinterrupted — Function

Test if the Job was interrupted during running.

EasyJobs.pendingjobs — Function

pendingjobs(jobs)

Filter only the pending jobs in a sequence of Jobs.

EasyJobs.runningjobs — Function

runningjobs(jobs)

Filter only the running jobs in a sequence of Jobs.

EasyJobs.exitedjobs — Function

exitedjobs(jobs)

Filter only the exited jobs in a sequence of Jobs.

EasyJobs.succeededjobs — Function

succeededjobs(jobs)

Filter only the succeeded jobs in a sequence of Jobs.

EasyJobs.failedjobs — Function

failedjobs(jobs)

Filter only the failed jobs in a sequence of Jobs.

EasyJobs.interruptedjobs — Function

interruptedjobs(jobs)

Filter only the interrupted jobs in a sequence of Jobs.

EasyJobs.createdtime — Function

Return the created time of a Job.

EasyJobs.starttime — Function

starttime(job::Job)

Return the start time of a Job. Return nothing if it is still pending.

EasyJobs.stoptime — Function

stoptime(job::Job)

Return the stop time of a Job. Return nothing if it has not exited.

EasyJobs.elapsed — Function

elapsed(job::Job)

Return the elapsed time of a Job since it started running.

If nothing, the Job is still pending. If it is finished, return how long it took to complete.

EasyJobs.description — Function

description(job::Job)

Return the description of a Job.

EasyJobs.run! — Method

run!(job::Job; n=1, δt=1, t=0)

Run a Job with maximum n attempts, with each attempt separated by δt seconds.

EasyJobs.interrupt! — Function

interrupt!(job::Job)

Manually interrupt a Job, works only if it is running.

EasyJobs.queue — Function

queue(; sortby = :created_time)

Print all Jobs that are pending, running, or finished as a table.

Accpetable arguments for sortby are :created_time, :user, :start_time, :stop_time, :elapsed, :status, and :times.

EasyJobs.query — Function

query(id::Integer)
query(ids::AbstractVector{<:Integer})

Query a specific (or a list of Jobs) by its (theirs) ID.

EasyJobs.ntimes — Function

ntimes(id::Integer)
ntimes(job::Job)

Return how many times a Job has been rerun.